+++ buession-beans/src/main/java/com/buession/beans/converters/LocalDatePropertyConverter.java
@@ -35,6 +35,7 @@ import java.util.TimeZone;
  * @author Yong.Teng
  * @since 2.3.1
  */
+@Deprecated
 public final class LocalDatePropertyConverter extends DateTimePropertyConverter<LocalDate> {
 
 	/**