+++ buession-beans/src/main/java/com/buession/beans/converters/SqlDatePropertyConverter.java
@@ -34,6 +34,7 @@ import java.util.TimeZone;
  * @author Yong.Teng
  * @since 2.3.1
  */
+@Deprecated
 public class SqlDatePropertyConverter extends DateTimePropertyConverter<Date> {
 
 	/**